fix: allow a range of dagger version
