+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/cheggaaa/pb v1.0.29
 	github.com/fatih/color v1.13.0
 	github.com/hashicorp/go-getter/v2 v2.0.1
+	github.com/hashicorp/go-version v1.2.0
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/moby/buildkit v0.10.1
 	github.com/otiai10/copy v1.7.0
@@ -52,7 +53,6 @@ require (
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
 	github.com/hashicorp/go-multierror v1.1.1 // indirect
 	github.com/hashicorp/go-safetemp v1.0.0 // indirect
-	github.com/hashicorp/go-version v1.2.0 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
 	github.com/imdario/mergo v0.3.12 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
