--- version/version.go
@@ -1,15 +1,0 @@
-package version
-
-const (
-	// DevelopmentVersion is unstable
-	DevelopmentVersion = "devel"
-)
-
-var (
-	// Version holds the complete version number. Filled in at linking time.
-	Version = DevelopmentVersion
-
-	// Revision is filled with the VCS (e.g. git) revision being used to build
-	// the program at linking time.
-	Revision = ""
-)
